Deleted Lua Scripting, Began Testing Engine Core
I deleted all Lua scripting pieces within the code as I have decided that it does not fit within the scope of this project for the very foreseeable future. Could be revisited one day maybe, kind of doubtful. I also wrote a simple test for the GameLoop to build a black screen which worked. Now I can move forward with further testing