Animation Rendering Test Passed
Loaded in animations for a newly created player gameobject. Then i was able to successfully get it to render to screen and animate. I also moved the Swing setup into the GameWindow class, so that I could get rid of some of the abstraction in the TestGame. Now it feels more plug and play.