Reduced Boilerplate
Made some simple hooks, so that it is easier to quickly load things in. Reducing the boilerplate needed to test way smaller. You can still go the longer way, and may need to when building a full game, but for now this should at least help me with test, and others for making a small Demo. Will look into further modularity later...