FINAL UPDATE FOR FIRST ALPHA!
Made the engine more of an engine. I built a lot of stuff in a way that was more game like, so I went back and made it more modular. This Involved deleting the Player entity, now everything will just come off of GameObject. I think there is a way to do it better but for now at least, I want the engine to be just a little more open ended. I also took out all the default methods. It was fine for quick testing, but I want users to have to have more intent when using the engine.

I also created Startable and Updatable. Now people can actually do something like create a script for their game called PlayerMovement and have it implement Startable and Updatable, that way it works more engine like, more Unity like in a sense. The user has to call that themselves. Unlike Unity, there is no MonoBehaviour. If your script needs to update every frame, implement it as Updatable.

Also updated the AnimationLoader to load in batches of animations with names, so it is easier for the user to call it in the AnimationController.

Speaking of I added in an AnimationController. Now it is a little more intuitive to change animaitons for whatever GameObject it is connected to.

The camera was rebuilt completely, and even though I am still not entirely happy with it now, it will have to work for the time being. But I would not be surprised to see me go back through and completely rebuild it again. It is just pretty messy and buggy.

Also I went through and tried my best to make everything Pixel based, so things would draw and render and animate based on the spite sizing. I did it in an effort to make the camera work and it sort of helped, but it is in my opinion a generally good idea to go ahead and do this now since the engine is aiming to be Pixel based.

And as always I updated all of the documentation, I tried to make it easy to follow for everyone.

Cheers, Case